listar vehiculos y cambio comprobarAdmin
Listar vehículos disponibles y no disponibles hecho. El metodo comprobarAdmin se ha cambiado por comprobarUser y devuelve el tipo de usuario que introduzcas.

+++ ISO2023/src/main/java/edu/uclm/esi/iso/ISO2023/services/ClienteService.java
@@ -17,6 +17,8 @@ import edu.uclm.esi.iso.ISO2023.exceptions.*;
 @Service
 public class ClienteService {
 	@Autowired
+	private UserService userService;
+	@Autowired
 	private ClienteDAO clienteDAO;
 	@Autowired
 	private SeguridadService comprobarSeguridad;
@@ -24,50 +26,56 @@ public class ClienteService {
 	private AdminService adminService;
 
 	public List<Cliente> listaClientes(String emailAdmin, String passwordAdmin) {
-		adminService.comprobarAdmin(emailAdmin, passwordAdmin);
-		return this.clienteDAO.findAll();
+		if (userService.comprobarUsuario(emailAdmin, passwordAdmin).equals("admin"))
+			return this.clienteDAO.findAll();
+		else {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "No tienes permisos para realizar esta accion.");
+		}
 	}
 
 	public void actualizarCliente(String nombre, String apellidos, String email, String password, boolean activo,
-			int intentos, String fechaNacimiento, String carnet, String telefono, String dni, String emailAdmin, String passwordAdmin)
-			throws contraseñaIncorrecta, formatoIncompleto, numeroInvalido {
-		adminService.comprobarAdmin(emailAdmin, passwordAdmin);
-		Cliente cliente = new Cliente(nombre, apellidos, email, password, activo, intentos, fechaNacimiento, carnet,telefono, dni);
-		
-		Optional<Cliente> clienteExiste = clienteDAO.findByEmail(email);
-		if (clienteExiste.isPresent()) {
-			comprobarSeguridad.restriccionesPassword(cliente);
-			comprobarSeguridad.validarEmail(cliente.getEmail());
-			comprobarSeguridad.comprobarNumero(cliente.getTelefono());
-	
-			if (Boolean.FALSE.equals(comprobarSeguridad.comprobarDni(cliente.getDni())))
-				throw new numeroInvalido(
-						"El NIF introducido no es un NIF valido. Tiene que contener 8 numeros y un caracter");
-	
-			if (cliente.getPassword().length() != 60) {
-				cliente.setPassword(comprobarSeguridad.cifrarPassword(cliente.getPassword()));
+			int intentos, String fechaNacimiento, String carnet, String telefono, String dni, String emailAdmin,
+			String passwordAdmin) throws contraseñaIncorrecta, formatoIncompleto, numeroInvalido {
+		if (userService.comprobarUsuario(emailAdmin, passwordAdmin).equals("admin")) {
+			Cliente cliente = new Cliente(nombre, apellidos, email, password, activo, intentos, fechaNacimiento, carnet,
+					telefono, dni);
+
+			Optional<Cliente> clienteExiste = clienteDAO.findByEmail(email);
+			if (clienteExiste.isPresent()) {
+				comprobarSeguridad.restriccionesPassword(cliente);
+				comprobarSeguridad.validarEmail(cliente.getEmail());
+				comprobarSeguridad.comprobarNumero(cliente.getTelefono());
+
+				if (Boolean.FALSE.equals(comprobarSeguridad.comprobarDni(cliente.getDni())))
+					throw new numeroInvalido(
+							"El NIF introducido no es un NIF valido. Tiene que contener 8 numeros y un caracter");
+
+				if (cliente.getPassword().length() != 60) {
+					cliente.setPassword(comprobarSeguridad.cifrarPassword(cliente.getPassword()));
+				}
+				cliente.setNombre(nombre);
+				cliente.setApellidos(apellidos);
+				cliente.setEmail(email);
+				cliente.setPassword(password);
+				cliente.setActivo(activo);
+				cliente.setIntentos(intentos);
+				cliente.setFechaNacimiento(fechaNacimiento);
+				cliente.setCarnet(carnet);
+				cliente.setTelefono(telefono);
+				cliente.setDni(dni);
+				clienteDAO.save(cliente);
 			}
-			cliente.setNombre(nombre);
-			cliente.setApellidos(apellidos);
-			cliente.setEmail(email);
-			cliente.setPassword(password);
-			cliente.setActivo(activo);
-			cliente.setIntentos(intentos);
-			cliente.setFechaNacimiento(fechaNacimiento);
-			cliente.setCarnet(carnet);
-			cliente.setTelefono(telefono);
-			cliente.setDni(dni);
-			clienteDAO.save(cliente);
 		}
 	}
-	
+
 	public void eliminarCliente(String email, String emailAdmin, String passwordAdmin) {
-		adminService.comprobarAdmin(emailAdmin, passwordAdmin);
-		Optional<Cliente> clienteExiste = clienteDAO.findByEmail(email);
-		if (clienteExiste.isPresent()) {
-			clienteDAO.deleteById(email);
-		} else {
-			throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Ese cliente no existe");
+		if (userService.comprobarUsuario(emailAdmin, passwordAdmin).equals("admin")) {
+			Optional<Cliente> clienteExiste = clienteDAO.findByEmail(email);
+			if (clienteExiste.isPresent()) {
+				clienteDAO.deleteById(email);
+			} else {
+				throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Ese cliente no existe");
+			}
 		}
 	}
 }
